[MOV] spreadsheet: move some tests from odoo to spreadsheet
Task: 4074487
